Optimize a description Status's Reason
Optimized the display of scheduling failure reasons. There are many labels
in the cluster, so you can go directly to managedcluster to view them if
necessary, and this information will no longer be displayed when the scheduling
fails. At the same time, the reason for the failure is updated to the
description status's 'Reason', which is convenient for the User System to query
and display.